# Pool buffers for quoted strings and byte slices