--- a/jsonstring.go
+++ b/jsonstring.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"io"
 	"strconv"
+	"sync"
 	"unicode"
 	"unicode/utf16"
 	"unicode/utf8"
@@ -17,9 +18,24 @@ import (
 
 var hex = "0123456789abcdef"
 
+var bufferPool = sync.Pool{
+	New: func() interface{} {
+		return &bytes.Buffer{}
+	},
+}
+
+func getBuffer() *bytes.Buffer {
+	return bufferPool.Get().(*bytes.Buffer)
+}
+
+func poolBuffer(buf *bytes.Buffer) {
+	buf.Reset()
+	bufferPool.Put(buf)
+}
+
 // NOTE: keep in sync with writeQuotedBytes below.
 func writeQuotedString(w io.Writer, s string) (int, error) {
-	buf := &bytes.Buffer{}
+	buf := getBuffer()
 	buf.WriteByte('"')
 	start := 0
 	for i := 0; i < len(s); {
@@ -70,12 +86,14 @@ func writeQuotedString(w io.Writer, s string) (int, error) {
 		buf.WriteString(s[start:])
 	}
 	buf.WriteByte('"')
-	return w.Write(buf.Bytes())
+	n, err := w.Write(buf.Bytes())
+	poolBuffer(buf)
+	return n, err
 }
 
 // NOTE: keep in sync with writeQuoteString above.
 func writeQuotedBytes(w io.Writer, s []byte) (int, error) {
-	buf := &bytes.Buffer{}
+	buf := getBuffer()
 	buf.WriteByte('"')
 	start := 0
 	for i := 0; i < len(s); {
@@ -126,7 +144,9 @@ func writeQuotedBytes(w io.Writer, s []byte) (int, error) {
 		buf.Write(s[start:])
 	}
 	buf.WriteByte('"')
-	return w.Write(buf.Bytes())
+	n, err := w.Write(buf.Bytes())
+	poolBuffer(buf)
+	return n, err
 }
 
 // getu4 decodes \uXXXX from the beginning of s, returning the hex value,