@ChrisHines 2016-06-01
Thanks for the contribution. This code is not compatible with Go 1.2, which caused the test failure. Please remove Go 1.2 from the .travis.yml file and update this PR then it should be good to merge.